fix(tests): separate desktop inspection specs
