+++ apps/desktop/e2e/live-work-rail-sticky-gap.spec.ts
@@ -80,55 +80,3 @@ test("LiveWorkRail sticky edit header and file-toggle pin without a gap", async 
     await app.close();
   }
 });
-
-// Inspect-style spec that captures a screenshot for visual review.
-// Gated behind the same PWRAGENT_SCREENSHOT_CAPTURE env var the other
-// inspect specs use; not run in CI.
-test.describe("inspect", () => {
-  test.skip(
-    !process.env.PWRAGENT_RAIL_INSPECT,
-    "Set PWRAGENT_RAIL_INSPECT=1 to capture screenshots locally.",
-  );
-
-  test("screenshot the rail with a long diff scrolled into the middle", async () => {
-    const app = await launchElectronApp({
-      fixturePath: path.resolve(
-        specDir,
-        "fixtures/live-work-rail-sticky-gap/replay.fixture.json"
-      ),
-      windowSize: { width: 1280, height: 720 },
-    });
-
-    try {
-      await app.window
-        .getByRole("button", { name: /LiveWorkRail sticky-gap replay/i })
-        .first()
-        .click();
-      await app.window
-        .getByLabel("Reply")
-        .fill("Drop in a long changelog entry to force rail-body scroll.");
-      await app.window.getByRole("button", { name: "Send" }).click();
-      await app.advance({ stepId: "status-active-1" });
-      await app.advance({ stepId: "turn-started-1" });
-      await app.advance({ stepId: "turn-diff-updated-1" });
-
-      const rail = app.window.getByRole("complementary", {
-        name: /Edited 1 file/i,
-      });
-      await rail.getByRole("button", { name: /Update CHANGELOG\.md/i }).click();
-      const railBody = rail.locator("css=.live-work-rail__body");
-      await railBody.evaluate((el) => {
-        el.scrollTop = 200;
-      });
-
-      await app.window.screenshot({
-        path: path.resolve(
-          specDir,
-          "../test-results/live-work-rail-sticky-gap.png",
-        ),
-      });
-    } finally {
-      await app.close();
-    }
-  });
-});
