+++ apps/desktop/package.json
@@ -43,12 +43,13 @@
     "eval:typecheck": "tsc --noEmit -p eval/tsconfig.json",
     "preview": "electron-vite preview",
     "test:e2e": "playwright test -c playwright.config.ts",
-    "inspect:e2e:branch-drift": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_E2E_INSPECT=1 playwright test -c playwright.config.ts e2e/thread-branch-drift.inspect.spec.ts --headed --timeout=0",
-    "inspect:e2e:workspace-handoff": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_E2E_INSPECT=1 playwright test -c playwright.config.ts e2e/workspace-handoff-dialog.inspect.spec.ts --headed --timeout=0",
-    "inspect:e2e:markdown-table": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_E2E_INSPECT=1 playwright test -c playwright.config.ts e2e/markdown-findings-table.inspect.spec.ts --headed --timeout=0",
-    "inspect:e2e:update-banner": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_E2E_INSPECT=1 playwright test -c playwright.config.ts e2e/update-banner.inspect.spec.ts --headed --timeout=0",
-    "screenshot:readme": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_SCREENSHOT_CAPTURE=1 playwright test -c playwright.config.ts e2e/readme-screenshots.inspect.spec.ts --headed && node scripts/filter-noise-screenshots.mjs",
-    "screenshot:docs-site": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_DOCS_SITE_SCREENSHOT_CAPTURE=1 playwright test -c playwright.config.ts e2e/docs-site-screenshots.inspect.spec.ts --headed && node scripts/filter-noise-screenshots.mjs --root \"${PWRAGENT_DOCS_SITE_REPO:-$HOME/github/docs.pwragent.ai}\"",
+    "inspect:e2e:branch-drift": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_E2E_INSPECT=1 playwright test -c playwright.inspect.config.ts e2e/thread-branch-drift.inspect.spec.ts --headed --timeout=0",
+    "inspect:e2e:workspace-handoff": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_E2E_INSPECT=1 playwright test -c playwright.inspect.config.ts e2e/workspace-handoff-dialog.inspect.spec.ts --headed --timeout=0",
+    "inspect:e2e:markdown-table": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_E2E_INSPECT=1 playwright test -c playwright.inspect.config.ts e2e/markdown-findings-table.inspect.spec.ts --headed --timeout=0",
+    "inspect:e2e:update-banner": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_E2E_INSPECT=1 playwright test -c playwright.inspect.config.ts e2e/update-banner.inspect.spec.ts --headed --timeout=0",
+    "inspect:e2e:live-work-rail-sticky-gap": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_RAIL_INSPECT=1 playwright test -c playwright.inspect.config.ts e2e/live-work-rail-sticky-gap.inspect.spec.ts --headed",
+    "screenshot:readme": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_SCREENSHOT_CAPTURE=1 playwright test -c playwright.inspect.config.ts e2e/readme-screenshots.inspect.spec.ts --headed && node scripts/filter-noise-screenshots.mjs",
+    "screenshot:docs-site": "node scripts/ensure-electron-runtime.mjs && node scripts/rebuild-native-for-electron.mjs && pnpm build && PWRAGENT_DOCS_SITE_SCREENSHOT_CAPTURE=1 playwright test -c playwright.inspect.config.ts e2e/docs-site-screenshots.inspect.spec.ts --headed && node scripts/filter-noise-screenshots.mjs --root \"${PWRAGENT_DOCS_SITE_REPO:-$HOME/github/docs.pwragent.ai}\"",
     "typecheck": "tsc --noEmit -p tsconfig.json"
   },
   "dependencies": {
